convert sign out SA to a public action

+++ src/app/[lng]/(auth)/log-ud/actions.ts
@@ -1,12 +1,14 @@
 'use server'
 
-import { privateAction } from '@/lib/safe-action'
+import { publicAction } from '@/lib/safe-action'
 import { sessionService } from '@/service/session'
 import { redirect } from 'next/navigation'
 
-export const signOutAction = privateAction
+export const signOutAction = publicAction
   .metadata({ actionName: 'signOut' })
   .action(async ({ ctx }) => {
-    await sessionService.delete(ctx.session.id)
+    if (ctx.session) {
+      await sessionService.delete(ctx.session.id)
+    }
     return redirect(`/${ctx.lang}/log-ind`)
   })
